# Improve status message access