--- a/utils_test.go
+++ b/utils_test.go
@@ -179,12 +179,6 @@ func Test_Utils_StatusMessage(t *testing.T) {
 	res := StatusMessage(204)
 	AssertEqual(t, "No Content", res)
 
-	res = StatusMessage(-1)
-	AssertEqual(t, "", res)
-
-	res = StatusMessage(0)
-	AssertEqual(t, "", res)
-
 	res = StatusMessage(404)
 	AssertEqual(t, "Not Found", res)
 
@@ -195,6 +189,15 @@ func Test_Utils_StatusMessage(t *testing.T) {
 	AssertEqual(t, "Network Authentication Required", res)
 
 	res = StatusMessage(1337)
+	AssertEqual(t, "", res)
+
+	res = StatusMessage(-1)
+	AssertEqual(t, "", res)
+
+	res = StatusMessage(0)
+	AssertEqual(t, "", res)
+
+	res = StatusMessage(600)
 	AssertEqual(t, "", res)
 }