```go
// new
Benchmark_StatusMessage-12        1000000000             0.669 ns/op           0 B/op           0 allocs/op
Benchmark_StatusMessage-12        1000000000             0.678 ns/op           0 B/op           0 allocs/op
Benchmark_StatusMessage-12        1000000000             0.650 ns/op           0 B/op           0 allocs/op
// old
Benchmark_StatusMessage-12        160494153             7.80 ns/op           0 B/op           0 allocs/op
Benchmark_StatusMessage-12        156866020             7.67 ns/op           0 B/op           0 allocs/op
Benchmark_StatusMessage-12        156083289             7.74 ns/op           0 B/op           0 allocs/op
```